Rewrite ReleaseCommand.
